requirements.txt
dash==1.8.0
dash_canvas==0.1.0
dash_core_components==1.7.0
dash_daq==0.5.0
dash_html_components==1.0.2
dash_table==4.6.0
numpy==1.19.5
pandas==1.1.5
Pillow==9.0.0
tensorflow-cpu==2.0.0
